+++ src/javascript/components/plugins/theme_card.js
@@ -26,7 +26,7 @@ const theme_card = puffin.element(`
         function setTheme(){
             graviton.setTheme(this.getAttribute("name"))
             selectionFromTo(this.parentElement,this);
-            graviton.saveConfiguration()
+            GravitonState.emit("ConfigurationChanged")
         }
     ]
 });